[client] Rename MessageProducer classes to Producer
